% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_details.R
\name{tiktokcomm_get_ad_details}
\alias{tiktokcomm_get_ad_details}
\title{Get TikTok Ad Details}
\usage{
tiktokcomm_get_ad_details(
  ad_id,
  fields = c("ad.id", "ad.first_shown_date", "ad.last_shown_date", "ad.status",
    "ad.status_statement", "ad.videos", "ad.image_urls", "ad.reach",
    "advertiser.business_id", "advertiser.business_name", "advertiser.paid_for_by",
    "advertiser.follower_count", "advertiser.avatar_url", "advertiser.profile_url",
    "ad_group.targeting_info"),
  verbose = FALSE
)
}
\arguments{
\item{ad_id}{Numeric. The ID of the ad to retrieve details for.}

\item{fields}{Character vector of fields to retrieve.
Default includes all available fields.}

\item{verbose}{Logical. If TRUE, provides detailed output. Default is FALSE.}
}
\value{
A tidy dataframe containing detailed information about the specified ad.
}
\description{
This function retrieves detailed information about a specific TikTok ad using the Commercial API
and returns the results as a tidy dataframe.
}
\examples{
\dontrun{
ad_details <- tiktokcomm_get_ad_details(
  ad_id = 104836593772645,
  verbose = TRUE
)
}
}
