% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_report.R
\name{tiktokcomm_get_ad_report}
\alias{tiktokcomm_get_ad_report}
\title{Get TikTok Ad Report}
\usage{
tiktokcomm_get_ad_report(
  start_date,
  end_date,
  country_code = "ALL",
  advertiser_ids = NULL,
  fields = "count_time_series_by_country",
  verbose = FALSE
)
}
\arguments{
\item{start_date}{Character. Start date for the report in "YYYY-MM-DD" format. Must be after 2022-10-01.}

\item{end_date}{Character. End date for the report in "YYYY-MM-DD" format.}

\item{country_code}{Character. The country code to filter by. Default is "ALL".}

\item{advertiser_ids}{Character vector or single character value. The advertiser business ID(s) to filter by.}

\item{fields}{Character vector. The fields to retrieve. Default is "count_time_series_by_country".}

\item{verbose}{Logical. If TRUE, provides detailed output. Default is FALSE.}
}
\value{
A tidy dataframe containing the ad report data.
}
\description{
This function retrieves a report on ad publishing from the TikTok Commercial API.
}
\examples{
\dontrun{
ad_report <- tiktokcomm_get_ad_report(
  start_date = "2023-01-02",
  end_date = "2023-01-09",
  country_code = "DE",
  advertiser_ids = "7057157514558702338",
  verbose = TRUE
)
}
}
