% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{tiktokcomm_get_token}
\alias{tiktokcomm_get_token}
\title{Retrieve TikTok Commercial API Token}
\usage{
tiktokcomm_get_token(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. If TRUE, provides detailed output. Default is FALSE.}
}
\value{
A list containing the client access token, expiration time, and token type.
}
\description{
This function retrieves the stored access token or authenticates if necessary.
}
\examples{
\dontrun{
token <- tiktokcomm_get_token()
}
}
