% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commercial.R
\name{tiktokcomm_query_commercial_content}
\alias{tiktokcomm_query_commercial_content}
\title{Query TikTok Commercial Content}
\usage{
tiktokcomm_query_commercial_content(
  start_date,
  end_date,
  creator_country_code = "ALL",
  creator_usernames = NULL,
  fields = c("id", "create_timestamp", "create_date", "label", "brand_names", "creator",
    "videos"),
  max_count = 10,
  max_pages = Inf,
  verbose = FALSE
)
}
\arguments{
\item{start_date}{Character. Start date for the query in "YYYY-MM-DD" format. Must be after 2022-10-01.}

\item{end_date}{Character. End date for the query in "YYYY-MM-DD" format.}

\item{creator_country_code}{Character. The country code of the content creator. Default is "ALL".}

\item{creator_usernames}{Character vector. The usernames of specific creators to query.}

\item{fields}{Character vector. The fields to retrieve. Default includes all available fields.}

\item{max_count}{Numeric. Maximum number of results to return per page (default 10, max 50).}

\item{max_pages}{Numeric. Maximum number of pages to retrieve. Default is Inf (all pages).}

\item{verbose}{Logical. If TRUE, provides detailed output. Default is FALSE.}
}
\value{
A tidy dataframe containing the queried commercial content data.
}
\description{
This function queries the TikTok Commercial API for commercial content based on specified criteria.
}
\examples{
\dontrun{
commercial_content <- tiktokcomm_query_commercial_content(
  start_date = "2023-01-02",
  end_date = "2023-01-09",
  creator_country_code = "FR",
  max_count = 20,
  verbose = TRUE
)
}
}
